+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/account/history/fixture/TransactionHistorySyncStateFixture.kt
@@ -3,8 +3,8 @@ package co.electriccoin.zcash.ui.screen.account.history.fixture
 import cash.z.ecc.android.sdk.fixture.TransactionOverviewFixture
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.TransactionRecipient
+import co.electriccoin.zcash.ui.screen.account.ext.TransactionOverviewExt
 import co.electriccoin.zcash.ui.screen.account.state.TransactionHistorySyncState
-import co.electriccoin.zcash.ui.screen.account.state.TransactionOverviewExt
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf