+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/state/TransactionHistorySyncState.kt
@@ -1,21 +1,14 @@
 package co.electriccoin.zcash.ui.screen.account.state
 
+import co.electriccoin.zcash.ui.screen.account.ext.TransactionOverviewExt
 import kotlinx.collections.immutable.ImmutableList
 
-sealed class TransactionHistorySyncState {
-    object Loading : TransactionHistorySyncState() {
-        override fun toString() = "Loading" // NON-NLS
-    }
+sealed interface TransactionHistorySyncState {
+    data object Loading : TransactionHistorySyncState
 
-    data class Syncing(val transactions: ImmutableList<TransactionOverviewExt>) : TransactionHistorySyncState() {
-        fun hasNoTransactions(): Boolean {
-            return transactions.isEmpty()
-        }
-    }
+    sealed class Prepared(open val transactions: ImmutableList<TransactionOverviewExt>) : TransactionHistorySyncState
 
-    data class Done(val transactions: ImmutableList<TransactionOverviewExt>) : TransactionHistorySyncState() {
-        fun hasNoTransactions(): Boolean {
-            return transactions.isEmpty()
-        }
-    }
+    data class Syncing(override val transactions: ImmutableList<TransactionOverviewExt>) : Prepared(transactions)
+
+    data class Done(override val transactions: ImmutableList<TransactionOverviewExt>) : Prepared(transactions)
 }